feat(devtools): ship devtools as a layer (Nuxt SEO Model C)
