--- devtools/app.config.ts
@@ -1,27 +1,0 @@
-export default {
-  ui: {
-    colors: {
-      primary: 'green',
-      neutral: 'neutral',
-    },
-    button: {
-      defaultVariants: {
-        color: 'neutral',
-        variant: 'ghost',
-        size: 'sm',
-      },
-    },
-    badge: {
-      defaultVariants: {
-        color: 'neutral',
-        variant: 'subtle',
-        size: 'xs',
-      },
-    },
-    tooltip: {
-      defaultVariants: {
-        delayDuration: 0,
-      },
-    },
-  },
-}
