--- devtools/composables/rpc.ts
@@ -1,6 +1,0 @@
-import { useDevtoolsConnection } from 'nuxtseo-layer-devtools/composables/rpc'
-import { refreshSources } from './state'
-
-useDevtoolsConnection({
-  onConnected: () => refreshSources(),
-})
