+++ devtools/lib/sitemap/types.ts
@@ -1,0 +1,4 @@
+export type ModuleRuntimeConfig = any
+export type SitemapDefinition = any
+export type SitemapSourceResolved = any
+export type ProductionDebugResponse = any
