--- devtools/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@nuxtjs/sitemap-client",
-  "private": true,
-  "devDependencies": {
-    "@iconify-json/carbon": "catalog:",
-    "@iconify-json/simple-icons": "catalog:",
-    "@nuxt/devtools-kit": "catalog:",
-    "@nuxt/kit": "catalog:",
-    "@vueuse/core": "catalog:",
-    "nuxt": "catalog:",
-    "nuxtseo-layer-devtools": "catalog:",
-    "vue": "catalog:",
-    "vue-router": "catalog:"
-  }
-}
